refactor: move format_token to utils
